pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xgripper"
version = "0.0.1"
description = "Software SDK for Xense FlareGripper"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"

authors = [
  { name = "Jin Liu", email = "liujin@xenserobotics.com" },
  { name = "Hongzhan Yu", email = "yuhongzhan@xenserobotics.com" },
  { name = "Xincheng Yang", email = "yangxincheng@xenserobotics.com" },
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.10",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

keywords = ["xense robotics", "gripper", "data collection"]

dependencies = [
  "av",
  "matplotlib",
  "numpy",
  "opencv-python",
  "pyserial",
  "scipy",
  "xensesdk>=1.6.3",
  "pysurvive",
  "draccus",
  "pygame",
  "rerun-sdk",
  "spdlog",
]

[project.scripts]
xensegripper = "xensegripper.__main__:main"

[tool.setuptools]
include-package-data = true # include non .py files

[tool.setuptools.packages.find]
where = ["."]

# package data
# [tool.setuptools.package-data]
# xgripper = ["assets/**"]

# include non .py files
# [tool.setuptools.include-package-data]
# include-package-data = true
